Antispam for application process
Spambots have found the application form and I'm getting several spammy
applications a day. These spammy applications all have one thing in common: The
'submit' button is included in the output. This should never happen since the
'submit' button doesn't have a name and thus should not be sent by legitimate
browsers, so this change redirects those users to another HTML page. (And just
in case it does happen to be legitimate (which has never happened so far), the
page suggests using a different browser.)