Fix bad check in setup process
The COPYDHROOT variable was initially intended to be directly supplied
on the command line (and would be either "" or "1"), but before commit
was changed to be controlled by a switch instead. I forgot to update
this line to match, which meant that any Dreamhack created in this time
would have been set up incorrectly.