Get the reinstaller working again
Finally had the time to get this back up and running. This reinstaller
does NOT do anything with the GitHub repositories; it merely clones
whatever is there. We should have instructions somewhere for how to
delete and re-fork the repositories if someone wants a full reinstall
from scratch.

Later on I'll do the promised OAuth linkage to allow people to do this
from the Dreamhack server by running another command, but for now this
will do.