Map remote branches to local branches
When creating a new user or reinstalling, map the remote branches on
GitHub to local tracking branches automatically, except for "develop"
and "master" (which remain tracked to dreamwidth). This allows people to
jump straight back into working on a branch in progress without needing
to worry about setting up their branches first.