Leave /dreamhack/var/maintain.d/.placeholder alone
For ages I've tried to work out why this file was mysteriously
disappearing and subsequently causing git to delete the parent
directory, messing up its permissions in the process. Turns out it was
getting deleted by maintain-dreamhacks after it assumed it was a file to
take action on.

This change tells maintain-dreamhacks to leave the file alone - not to
take action on it and not to delete it.