dreamhacks_ro should not have INSERT access
I can't figure out why I ever stated that the read-only database user should
get INSERT access, but I've looked through the code and can't find a case where
it's necessary. Things *shouldn't* break with this change when it's actually
implemented on the server; if they do, that's a failure on my part.