Fix reinstaller for config file location
The move of the default configuration files from doc/ to etc/ break the
installer, so this fixes the issue. It also adds a README.txt file on install
that explains that changes made to that directory might not take effect.

No other changes are required as none of the other config changes clash with
lib/bin/do-config-stuff.