File modifications for new server
This commit has all the modifications needed for the new server. A lot of
outdated things have been fixed as I took the time to make sure everything was
good to go.

This commit doesn't include a new INSTALL file; I do however have notes for all
the changes necessary and will be fixing it soon.